+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file based on the
 
 ### Bugfixes
 
+* Fixed handling of Search::OPTION_SEARCH_IGNORE_UNAVAILABLE inside Scroll object
+
 ### Added
 
 ### Improvements